+++ packages/js-drive/lib/abci/handlers/errors/DataCorruptedError.js
@@ -1,8 +1,10 @@
 const DriveError = require('../../../errors/DriveError');
 
 class DataCorruptedError extends DriveError {
-  constructor() {
-    super('Data is corrupted. Please reset you node');
+  constructor(e) {
+    super('Cant\' commit previous block. Please reset you node');
+
+    this.error = e;
   }
 }