+++ packages/js-drive/test/unit/abci/handlers/stateTransition/unserializeStateTransitionFactory.spec.js
@@ -152,9 +152,9 @@ describe('unserializeStateTransitionFactory', () => {
       expect(loggerMock.info).to.have.been.calledOnceWithExactly(
         'Insufficient funds to process state transition',
       );
-      expect(loggerMock.debug).to.have.been.calledOnceWithExactly(
-        e,
-      );
+      expect(loggerMock.debug).to.have.been.calledOnceWithExactly({
+        consensusErrors: [error],
+      });
     }
   });
 });